Revert "reduce bundle size"
This reverts commit 4d09a2e9506a8f7a509390db16c47c27ba64194f.
