+++ packages/material-ui/src/InputBase/Textarea.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import useForkRef from '../utils/useForkRef';
+import { useForkRef } from '../utils/reactHelpers';
 import debounce from 'debounce'; // < 1kb payload overhead when lodash/debounce is > 3kb.
 
 const styles = {
@@ -20,7 +20,10 @@ function getStyleValue(computedStyle, property) {
   return parseInt(computedStyle[property], 10) || 0;
 }
 
-const useEnhancedEffect = typeof window !== 'undefined' ? React.useLayoutEffect : React.useEffect;
+const useEnhancedEffect =
+  typeof window !== 'undefined' && process.env.NODE_ENV !== 'test'
+    ? React.useLayoutEffect
+    : React.useEffect;
 
 /**
  * @ignore - internal component.
@@ -33,12 +36,12 @@ const Textarea = React.forwardRef(function Textarea(props, ref) {
   const { current: isControlled } = React.useRef(value != null);
   const inputRef = React.useRef();
   const [state, setState] = React.useState({});
-  const shallowRef = React.useRef();
+  const inputShallowRef = React.useRef();
   const handleRef = useForkRef(ref, inputRef);
 
-  const syncHeight = () => {
+  const syncHeightWithShadow = () => {
     const input = inputRef.current;
-    const inputShallow = shallowRef.current;
+    const inputShallow = inputShallowRef.current;
 
     const computedStyle = window.getComputedStyle(input);
     inputShallow.style.width = computedStyle.width;
@@ -85,7 +88,7 @@ const Textarea = React.forwardRef(function Textarea(props, ref) {
 
   React.useEffect(() => {
     const handleResize = debounce(() => {
-      syncHeight();
+      syncHeightWithShadow();
     }, 166); // Corresponds to 10 frames at 60 Hz.
 
     window.addEventListener('resize', handleResize);
@@ -96,12 +99,12 @@ const Textarea = React.forwardRef(function Textarea(props, ref) {
   });
 
   useEnhancedEffect(() => {
-    syncHeight();
+    syncHeightWithShadow();
   });
 
   const handleChange = event => {
     if (!isControlled) {
-      syncHeight();
+      syncHeightWithShadow();
     }
 
     if (onChange) {
@@ -126,7 +129,7 @@ const Textarea = React.forwardRef(function Textarea(props, ref) {
         aria-hidden="true"
         className={props.className}
         readOnly
-        ref={shallowRef}
+        ref={inputShallowRef}
         tabIndex={-1}
         style={styles.shadow}
       />
