+++ packages/material-ui/src/Popper/Popper.js
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types';
 import PopperJS from 'popper.js';
 import { chainPropTypes } from '@material-ui/utils';
 import Portal from '../Portal';
-import setRef from '../utils/setRef';
-import withForwardedRef from '../utils/withForwardedRef';
+import { setRef, withForwardedRef } from '../utils';
 
 function flipPlacement(placement) {
   const direction = (typeof window !== 'undefined' && document.body.getAttribute('dir')) || 'ltr';
