+++ packages/material-ui/src/Slide/Slide.js
@@ -6,7 +6,7 @@ import ReactDOM from 'react-dom';
 import EventListener from 'react-event-listener';
 import debounce from 'debounce'; // < 1kb payload overhead when lodash/debounce is > 3kb.
 import Transition from 'react-transition-group/Transition';
-import setRef from '../utils/setRef';
+import { setRef } from '../utils/reactHelpers';
 import withTheme from '../styles/withTheme';
 import { duration } from '../styles/transitions';
 import { reflow, getTransitionProps } from '../transitions/utils';
