+++ packages/material-ui/src/utils/index.js
@@ -1,9 +1,8 @@
 export { default as deprecatedPropType } from './deprecatedPropType';
 export * from './helpers';
 export * from './reactHelpers';
+export { default as requirePropFactory } from './requirePropFactory';
 export { default as ownerDocument } from './ownerDocument';
 export { default as ownerWindow } from './ownerWindow';
-export { default as requirePropFactory } from './requirePropFactory';
-export { default as setRef } from './setRef';
 export { default as unsupportedProp } from './unsupportedProp';
 export { default as withForwardedRef } from './withForwardedRef';
