--- packages/material-ui/src/utils/setRef.js
@@ -1,7 +1,0 @@
-export default function setRef(ref, value) {
-  if (typeof ref === 'function') {
-    ref(value);
-  } else if (ref) {
-    ref.current = value;
-  }
-}
